Extract shared ADO auth helpers to adoAuth.ts
